requirements.txt
asgiref==3.5.0
Django==4.0.4
django-rest-auth==0.9.5
djangorestframework==3.13.1
importlib-metadata==4.11.3
Markdown==3.3.6
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
zipp==3.8.0
